requirements.txt
Flask==0.10.1
Flask-Cors==2.0.0
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.4
cffi==0.9.2
characteristic==14.3.0
cryptography==0.8.1
cssselect==0.9.1
enum34==1.0.4
itsdangerous==0.24
lxml==3.4.2
ndg-httpsclient==0.3.3
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pymongo==2.8
python-dateutil==2.4.2
python-firebase==1.2
requests==2.6.0
service-identity==14.0.0
six==1.9.0
wsgiref==0.1.2
zope.interface==4.1.2
